Refactor: Extract `confirm` into a new mod
